+++ provider/cmd/pulumi-gen-kubernetes/main.go
@@ -29,7 +29,6 @@ import (
 	"text/template"
 	"unicode"
 
-	"github.com/pkg/errors"
 	"github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/gen"
 	providerVersion "github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/version"
 	"github.com/pulumi/pulumi/pkg/v3/codegen"
@@ -175,6 +174,7 @@ func generateSchema(swaggerPath string) schema.PackageSpec {
 // This is to mostly filter resources from the spec.
 var resourcesToFilterFromTemplate = codegen.NewStringSet(
 	"kubernetes:helm.sh/v3:Release",
+	"kubernetes:helm.sh/v4:Chart",
 	"kubernetes:yaml/v2:ConfigFile",
 	"kubernetes:yaml/v2:ConfigGroup",
 )
@@ -520,7 +520,7 @@ func makeJSONString(v any) ([]byte, error) {
 func mustWritePulumiSchema(pkgSpec schema.PackageSpec, version string) {
 	schemaJSON, err := makeJSONString(pkgSpec)
 	if err != nil {
-		panic(errors.Wrap(err, "marshaling Pulumi schema"))
+		panic(fmt.Errorf("marshaling Pulumi schema: %w", err))
 	}
 
 	mustWriteFile(BaseDir, filepath.Join("provider", "cmd", "pulumi-resource-kubernetes", "schema.json"), schemaJSON)
@@ -529,7 +529,7 @@ func mustWritePulumiSchema(pkgSpec schema.PackageSpec, version string) {
 	versionedPkgSpec.Version = version
 	versionedSchemaJSON, err := makeJSONString(versionedPkgSpec)
 	if err != nil {
-		panic(errors.Wrap(err, "marshaling Pulumi schema"))
+		panic(fmt.Errorf("marshaling Pulumi schema: %w", err))
 	}
 	mustWriteFile(BaseDir, filepath.Join("sdk", "schema", "schema.json"), versionedSchemaJSON)
 }