+++ provider/pkg/provider/yaml/v2/yaml.go
@@ -29,7 +29,6 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/pkg/errors"
 	"github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/clients"
 	"github.com/pulumi/pulumi-kubernetes/sdk/v4/go/kubernetes"
 	yamlv2 "github.com/pulumi/pulumi-kubernetes/sdk/v4/go/kubernetes/yaml/v2"
@@ -62,12 +61,12 @@ func Parse(ctx context.Context, opts ParseOptions) ([]unstructured.Unstructured,
 			// If the string looks like a URL, in that it begins with a scheme, fetch it over the network.
 			resp, err := http.Get(file)
 			if err != nil {
-				return nil, errors.Wrapf(err, "fetching YAML over network")
+				return nil, fmt.Errorf("fetching YAML over network: %w", err)
 			}
 			defer resp.Body.Close()
 			yaml, err = io.ReadAll(resp.Body)
 			if err != nil {
-				return nil, errors.Wrapf(err, "reading YAML over network")
+				return nil, fmt.Errorf("reading YAML over network: %w", err)
 			}
 			yamls = append(yamls, string(yaml))
 		} else {
@@ -77,7 +76,7 @@ func Parse(ctx context.Context, opts ParseOptions) ([]unstructured.Unstructured,
 			if opts.Glob && isGlobPattern(file) {
 				files, err = filepath.Glob(file)
 				if err != nil {
-					return nil, errors.Wrapf(err, "expanding glob")
+					return nil, fmt.Errorf("expanding glob: %w", err)
 				}
 				sort.Strings(files)
 			} else {
@@ -86,7 +85,7 @@ func Parse(ctx context.Context, opts ParseOptions) ([]unstructured.Unstructured,
 			for _, f := range files {
 				yaml, err = os.ReadFile(f)
 				if err != nil {
-					return nil, errors.Wrapf(err, "reading YAML file from disk")
+					return nil, fmt.Errorf("reading YAML file from disk: %w", err)
 				}
 				yamls = append(yamls, string(yaml))
 			}
@@ -100,7 +99,7 @@ func Parse(ctx context.Context, opts ParseOptions) ([]unstructured.Unstructured,
 		// Parse the resulting YAML bytes and turn them into raw Kubernetes objects.
 		dec, err := yamlDecode(yaml)
 		if err != nil {
-			return nil, errors.Wrapf(err, "decoding YAML")
+			return nil, fmt.Errorf("decoding YAML: %w", err)
 		}
 		objs = append(objs, dec...)
 	}
@@ -213,11 +212,15 @@ func expand(objs []unstructured.Unstructured) ([]unstructured.Unstructured, erro
 	return result, nil
 }
 
+type PreRegisterFunc func(ctx *pulumi.Context, apiVersion, kind, resourceName string, obj *unstructured.Unstructured,
+	resourceOpts []pulumi.ResourceOption) (*unstructured.Unstructured, []pulumi.ResourceOption)
+
 type RegisterOptions struct {
 	Objects         []unstructured.Unstructured
 	ResourcePrefix  string
 	SkipAwait       bool
 	ResourceOptions []pulumi.ResourceOption
+	PreRegisterF    PreRegisterFunc
 }
 
 // Register registers the given Kubernetes objects as resources with the Pulumi engine.
@@ -306,6 +309,13 @@ func register(
 			return nil, fmt.Errorf("YAML object is invalid: `%s`: %w", printUnstructured(obj), err)
 		}
 	}
+
+	if opts.PreRegisterF != nil {
+		obj, resourceOpts = opts.PreRegisterF(ctx, apiVersion, kind, resourceName, obj, resourceOpts)
+	}
+
+	// At this point, the object is a pure unstructured object (no inputs or outputs within it).
+	// Now it becomes an input to ctx.RegisterResource.
 
 	if fullKind == "v1/Secret" {
 		// Always mark these fields as secret to avoid leaking sensitive values from raw YAML.