+++ tests/gomega/kube.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2022, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,4 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package test
+package gomega
+
+import (
+	. "github.com/onsi/gomega/gstruct" //nolint:golint // gomega dsl
+	gomegatypes "github.com/onsi/gomega/types"
+)
+
+func HaveSkipAwaitAnnotation() gomegatypes.GomegaMatcher {
+	return MatchProps(IgnoreExtras, Props{
+		"metadata": MatchObject(IgnoreExtras, Props{
+			"annotations": MatchObject(IgnoreExtras, Props{
+				"pulumi.com/skipAwait": MatchValue("true"),
+			}),
+		}),
+	})
+}